LayoutTests/fast/scrolling/ios/bounding-client-rect-on-fixed-expected.txt

Verifies that getBoundingClientRect() on a fixedpos always returns the layout value.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS largestDelta is 0
PASS successfullyParsed is true

TEST COMPLETE
